updated test, code and corrections made to apis

+++ models/needModel.js
@@ -18,10 +18,11 @@ async function insertNeed(need) {
   .input('FileData', sql.VarBinary(sql.MAX), need.FileData)
   .input('FileName', sql.NVarChar(255), need.FileName)
   .input('extension', sql.NVarChar(10), need.extension)
-  .input('createdAt', sql.DateTime, need.createdAt);
+  .input('createdAt', sql.DateTime, need.createdAt)
+  .input('UserId', sql.Int, need.UserId);
   const result = await request.query(`
-    INSERT INTO NEED (NeedIs, Title, ContactPerson, Keywords, Proposal, Solution, FileData, FileName, extension, CreatedAt)
-    VALUES (@NeedIs, @Title, @ContactPerson, @Keywords, @Proposal, @Solution, @FileData, @FileName, @extension, @createdAt);
+    INSERT INTO NEED (NeedIs, Title, ContactPerson, Keywords, Proposal, Solution, FileData, FileName, extension, CreatedAt, UserId)
+    VALUES (@NeedIs, @Title, @ContactPerson, @Keywords, @Proposal, @Solution, @FileData, @FileName, @extension, @createdAt, @UserId);
     SELECT SCOPE_IDENTITY() AS id;
   `);
   await pool.close();
@@ -68,22 +69,25 @@ const updateNeed = async (id, updatedNeed) => {
     request.input('extension', sql.NVarChar(10), updatedNeed.extension);
     request.input('createdAt', sql.DateTime, updatedNeed.createdAt);
     request.input('id', sql.Int, id);
+    request.input('UserId', sql.Int, updatedNeed.UserId);
 
     const query = `UPDATE NEED SET NeedIs = @NeedIs, Title = @Title, ContactPerson = @ContactPerson,
       Keywords = @Keywords, Proposal = @Proposal, Solution = @Solution, FileData = @FileData,
-      FileName = @FileName, extension = @extension, createdAt = @createdAt WHERE id = @id`;
+      FileName = @FileName, extension = @extension, createdAt = @createdAt WHERE id = @id AND UserId = @UserId`;
+  
 
     const result = await request.query(query);
     await pool.close();
     return result.recordset;
 };
 
-async function deleteNeed(id) {
+async function deleteNeed(id, UserId) {
   try {
     const pool = await sql.connect(dbConfig);
     const request = pool.request();
-    const query = 'DELETE FROM NEED WHERE id = @id';
+    const query = 'DELETE FROM NEED WHERE id = @id AND UserId = @UserId'; 
     request.input('id', sql.Int, id);
+    request.input('UserId', sql.Int, UserId);
     const result = await request.query(query);
     return result.rowsAffected[0] > 0;
   } catch (error) {
@@ -93,6 +97,7 @@ async function deleteNeed(id) {
     sql.close();
   }
 }
+
 
 
 module.exports = {
